Adds community entity
This entity is responsible to hold a grid of plot fractions. For now it
is just a setup of the entity, without adding much of its behaviour. To
be added the stup of all plot fractions inside the creation of the
struct. Also, to be decided if this struct will include all the
households struct, that is how central will be this struct when running
the model.